Re-organise swap module
